requirements.txt
dnspython
pysocks
validators
flask_api
flask